Rename ContentFragment to EcTabFragment
Since EcFragment are in EcTabFragment, and content portal will have it's own viewpager, we'll rename the old ContentFragment to EcTabFragment
